requirements.txt
Flask==2.2.5
mysql-connector-python==8.2.0
pandas==1.5.3